ml/requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
catboost==1.2.7
joblib==1.4.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
